Add BASALISC paper to FHE hardware section
I see that the links in the **Hardware** section are to talks given, but since the BASALISC project is focused on hardware, I thought the link to the paper would be appropriate for this section. But if you'd like me to move it to the **Papers** section, I'll do that.

I'm working on the compiler and software stack at Niobium, one of the groups collaborating on BASALISC in the DARPA DPRIVE project.